+++ src/modules/voxelgenerator/lua/scripts/splitobjects.lua
@@ -1,21 +1,21 @@
 --
 -- moves all connected voxels of the node's volume into new volumes
 --
--- split all single identifiable objects into own layers/nodes
+-- split all single identifiable objects into own models/nodes
 --
 
 local vol = require "modules.volume"
 
 local splitVolume = {}
-local needNewLayer = false
+local needNewModel = false
 local palette = {}
 
 local function connectedVisitor(volume, x, y, z)
-	if needNewLayer then
-		local newLayer = scenegraph.new('splitobject', volume:region())
-		splitVolume = newLayer:volume()
-		newLayer:setPalette(palette)
-		needNewLayer = false
+	if needNewModel then
+		local newModel = scenegraph.new('splitobject', volume:region())
+		splitVolume = newModel:volume()
+		newModel:setPalette(palette)
+		needNewModel = false
 	end
 	local color = volume:voxel(x, y, z)
 	splitVolume:setVoxel(x, y, z, color)
@@ -26,7 +26,7 @@ function main(node, region, color)
 	palette = node:palette()
 	vol.visitYXZ(node:volume(), region, function (loopVolume, x, y, z)
 		if loopVolume:voxel(x, y, z) ~= -1 then
-			needNewLayer = true
+			needNewModel = true
 			splitVolume = nil
 			vol.visitConnected6(loopVolume, x, y, z, connectedVisitor)
 			if splitVolume ~= nil then