+++ src/modules/voxelutil/VolumeRescaler.h
@@ -121,8 +121,8 @@ void scaleDown(const SourceVolume& sourceVolume, const voxel::Palette &palette, 
 	// At this point the results are usable, but we have a problem with thin structures disappearing.
 	// For example, if we have a solid blue sphere with a one voxel thick layer of red voxels on it,
 	// then we don't care that the shape changes then the red voxels are lost but we do care that the
-	// color changes, as this is very noticable. Our solution is to process again only those voxels
-	// which lie on a material-air boundary, and to recompute their color using a larger naighbourhood
+	// color changes, as this is very noticeable. Our solution is to process again only those voxels
+	// which lie on a material-air boundary, and to recompute their color using a larger neighborhood
 	// while also accounting for how visible the child voxels are.
 	typename DestVolume::Sampler dstSampler(destVolume);
 	for (int32_t z = 0; z < destRegion.getDepthInVoxels(); ++z) {