+++ src/tools/voxedit/modules/voxedit-ui/SceneGraphPanel.h
@@ -41,7 +41,8 @@ private:
 public:
 	bool _popupNewModelNode = false;
 	bool init();
-	void update(video::Camera& camera, const char *title, ModelNodeSettings* layerSettings, command::CommandExecutionListener &listener);
+	void update(video::Camera &camera, const char *title, ModelNodeSettings *modelNodeSettings,
+				command::CommandExecutionListener &listener);
 	bool hasFocus() const;
 };